Move VerifySelection into base class because other beatmap might need the verify section.
