+++ osu.Game.Rulesets.Karaoke/Screens/Edit/Beatmaps/Lyrics/Settings/LyricEditorEditModeSection.cs
@@ -5,20 +5,11 @@
 
 using System;
 using osu.Framework.Allocation;
-using osu.Framework.Bindables;
-using osu.Framework.Graphics;
-using osu.Framework.Graphics.Containers;
-using osu.Framework.Graphics.Shapes;
 using osu.Framework.Localisation;
-using osu.Game.Graphics;
-using osu.Game.Graphics.Sprites;
-using osu.Game.Overlays.Toolbar;
-using osu.Game.Rulesets.Edit.Checks.Components;
 using osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Lyrics.Settings.Components.Markdown;
 using osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Lyrics.States;
 using osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Lyrics.States.Modes;
 using osu.Game.Rulesets.Karaoke.Screens.Edit.Components.Markdown;
-using osuTK.Graphics;
 
 namespace osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Lyrics.Settings
 {
@@ -57,91 +48,15 @@ namespace osu.Game.Rulesets.Karaoke.Screens.Edit.Beatmaps.Lyrics.Settings
             base.UpdateEditMode(mode);
         }
 
-        protected abstract partial class LyricEditorVerifySelection : Selection
+        protected abstract partial class LyricEditorVerifySelection : VerifySelection
         {
-            private readonly IBindableList<Issue> bindableIssues = new BindableList<Issue>();
-
-            protected LyricEditorVerifySelection()
-            {
-                CountCircle countCircle;
-
-                AddInternal(countCircle = new CountCircle
-                {
-                    Anchor = Anchor.TopRight,
-                    Origin = Anchor.Centre,
-                    X = -5,
-                });
-
-                bindableIssues.BindCollectionChanged((_, _) =>
-                {
-                    int count = bindableIssues.Count;
-                    countCircle.Alpha = count == 0 ? 0 : 1;
-                    countCircle.Count = count;
-                });
-            }
-
             [BackgroundDependencyLoader]
             private void load(ILyricEditorVerifier verifier)
             {
-                bindableIssues.BindTo(verifier.GetIssueByEditMode(EditMode));
+                Issues.BindTo(verifier.GetIssueByEditMode(EditMode));
             }
 
             protected abstract LyricEditorMode EditMode { get; }
-        }
-
-        /// <summary>
-        /// Copied from <see cref="ToolbarNotificationButton"/>
-        /// </summary>
-        private partial class CountCircle : CompositeDrawable
-        {
-            private readonly OsuSpriteText countText;
-            private readonly Circle circle;
-
-            private int count;
-
-            public int Count
-            {
-                get => count;
-                set
-                {
-                    if (count == value)
-                        return;
-
-                    if (value != count)
-                    {
-                        circle.FlashColour(Color4.White, 600, Easing.OutQuint);
-                        this.ScaleTo(1.1f).Then().ScaleTo(1, 600, Easing.OutElastic);
-                    }
-
-                    count = value;
-                    countText.Text = value.ToString("#,0");
-                }
-            }
-
-            public CountCircle()
-            {
-                AutoSizeAxes = Axes.X;
-                Height = 20;
-
-                InternalChildren = new Drawable[]
-                {
-                    circle = new Circle
-                    {
-                        RelativeSizeAxes = Axes.Both,
-                        Colour = Color4.Red
-                    },
-                    countText = new OsuSpriteText
-                    {
-                        Anchor = Anchor.Centre,
-                        Origin = Anchor.Centre,
-                        Y = -1,
-                        Font = OsuFont.GetFont(size: 18, weight: FontWeight.Bold),
-                        Padding = new MarginPadding(5),
-                        Colour = Color4.White,
-                        UseFullGlyphHeight = true,
-                    }
-                };
-            }
         }
     }
 }
